Fix an int->long conversion

+++ src/frontend/org/voltdb/DummyCommandLog.java
@@ -38,7 +38,7 @@ public class DummyCommandLog implements CommandLog {
     public void shutdown() throws InterruptedException {}
 
     @Override
-    public Semaphore logFault(Set<Integer> failedSites, Set<Long> faultedTxns) {
+    public Semaphore logFault(Set<Long> failedSites, Set<Long> faultedTxns) {
         return new Semaphore(1);
     }
